Refactoriza modelo con BigDecimal y equals hashCode

+++ src/Transferible.java
@@ -1,4 +1,9 @@
+import java.math.BigDecimal;
+
 public interface Transferible {
 
-    boolean transferir(CuentaBancaria destino, double monto);
+    boolean transferir(
+            CuentaBancaria destino,
+            BigDecimal monto
+    );
 }
